Reset branch tree to PR base snapshot

+++ projects/rccl/ext-profiler/example/event.h
@@ -49,7 +49,7 @@ struct netPlugin {
 };
 
 struct kernelCh {
-  uint64_t type;
+  uint8_t type;
   uint8_t channelId;
   struct taskEventBase* parent;
   double startTs;
